refactor: move pena accountability slice into core application
