+++ backend/src/api/dependencies/use_cases.py
@@ -13,6 +13,9 @@ from core.application.use_cases.get_season_match_insights_usecase import (
     GetSeasonMatchInsightsUseCase,
 )
 from core.application.use_cases.link_user_to_pena_usecase import LinkUserToPenaUseCase
+from core.application.use_cases.manage_pena_accountability_usecase import (
+    ManagePenaAccountabilityUseCase,
+)
 from core.application.use_cases.manage_pena_labels_usecase import (
     ManagePenaLabelsUseCase,
 )
@@ -23,9 +26,6 @@ from core.application.use_cases.update_player_profile_usecase import (
     UpdatePlayerProfileUseCase,
 )
 from fastapi import Depends
-from persistence.application.use_cases.manage_pena_accountability_usecase import (
-    ManagePenaAccountabilityUseCase,
-)
 from persistence.application.use_cases.manage_pena_membership_usecase import (
     ManagePenaMembershipUseCase,
 )
