+++ backend/src/api/middleware/exception_mapper.py
@@ -7,9 +7,14 @@ from typing import Any
 from auth.application.use_cases.login import InvalidCredentialsError
 from core.application.use_cases import (
     InvalidLinkTokenError,
+    InvalidPenaAccountabilityDataError,
     InvalidPenaLabelsDataError,
     InvalidPlayerUpdateDataError,
     PenaAccessDeniedError,
+    PenaAccountabilityAccessDeniedError,
+    PenaAccountabilityExpenseNotFoundError,
+    PenaAccountabilityMemberNotFoundError,
+    PenaAccountabilityPenaNotFoundError,
     PenaLabelsAccessDeniedError,
     PenaLabelsPenaNotFoundError,
     UserAlreadyLinkedError,
@@ -34,7 +39,6 @@ from fastapi import HTTPException, status
 from persistence.application.use_cases import (
     AdminUsernameExistsError,
     InvalidAdminRegistrationDataError,
-    InvalidPenaAccountabilityDataError,
     InvalidPenaGuestPlayerDataError,
     InvalidPenaMembershipUpdateDataError,
     InvalidPenaProfileImageError,
@@ -44,10 +48,6 @@ from persistence.application.use_cases import (
     InvalidSeasonMatchDataError,
     InvalidSeasonPlayerBatchDataError,
     InvalidSeasonPlayerUpdateDataError,
-    PenaAccountabilityAccessDeniedError,
-    PenaAccountabilityExpenseNotFoundError,
-    PenaAccountabilityMemberNotFoundError,
-    PenaAccountabilityPenaNotFoundError,
     PenaMembershipAccessDeniedError,
     PenaMembershipInvalidNationalityError,
     PenaMembershipNotFoundError,
