+++ backend/src/core/application/models/__init__.py
@@ -1,3 +1,11 @@
+from .pena_accountability_models import (
+    PenaAccountabilityExpenseCreate,
+    PenaAccountabilityExpenseInfo,
+    PenaAccountabilityInfo,
+    PenaAccountabilityMemberAccountInfo,
+    PenaAccountabilityMemberAccountUpsert,
+    PenaAccountabilitySettingsUpdate,
+)
 from .pena_labels_models import PenaLabelsInfo, PenaLabelsUpdate
 from .pena_link_models import PenaLinkToken
 from .pena_listing_models import PenaInfo, PenasPage, PenasPageResult, PenaSummary
@@ -16,6 +24,12 @@ __all__ = [
     "MatchInsightRow",
     "MatchPlayerStats",
     "MatchTeam",
+    "PenaAccountabilityExpenseCreate",
+    "PenaAccountabilityExpenseInfo",
+    "PenaAccountabilityInfo",
+    "PenaAccountabilityMemberAccountInfo",
+    "PenaAccountabilityMemberAccountUpsert",
+    "PenaAccountabilitySettingsUpdate",
     "PenaInfo",
     "PenaLinkToken",
     "PenaLabelsInfo",
