+++ backend/src/core/application/ports/__init__.py
@@ -1,4 +1,5 @@
 from .nationality_query_port import NationalityQueryPort
+from .pena_accountability_port import PenaAccountabilityPort
 from .pena_labels_port import PenaLabelsPort
 from .pena_link_port import PenaLinkPort
 from .pena_player_query_port import PenaPlayerQueryPort
@@ -8,9 +9,10 @@ from .player_profile_port import PlayerProfilePort
 from .season_match_insights_port import SeasonMatchInsightsPort
 
 __all__ = [
-    "PenaLinkPort",
     "NationalityQueryPort",
+    "PenaAccountabilityPort",
     "PenaLabelsPort",
+    "PenaLinkPort",
     "PenaPlayerQueryPort",
     "PenaProfilePort",
     "PenaQueryPort",
