+++ backend/src/core/application/use_cases/__init__.py
@@ -13,6 +13,14 @@ from .link_user_to_pena_usecase import (
     UserAlreadyLinkedError,
     UserProfileNotFoundError,
 )
+from .manage_pena_accountability_usecase import (
+    InvalidPenaAccountabilityDataError,
+    ManagePenaAccountabilityUseCase,
+    PenaAccountabilityAccessDeniedError,
+    PenaAccountabilityExpenseNotFoundError,
+    PenaAccountabilityMemberNotFoundError,
+    PenaAccountabilityPenaNotFoundError,
+)
 from .manage_pena_labels_usecase import (
     InvalidPenaLabelsDataError,
     ManagePenaLabelsUseCase,
@@ -46,6 +54,7 @@ __all__ = [
     "GetPlayerProfileUseCase",
     "GetSeasonMatchInsightsUseCase",
     "InvalidLinkTokenError",
+    "InvalidPenaAccountabilityDataError",
     "InvalidPenaLabelsDataError",
     "InvalidPenaProfileImageError",
     "InvalidNationalityError",
@@ -53,9 +62,14 @@ __all__ = [
     "InvalidProfileImageError",
     "InvalidSeasonInsightsDataError",
     "LinkUserToPenaUseCase",
+    "ManagePenaAccountabilityUseCase",
     "ManagePenaLabelsUseCase",
     "ManagePenaProfileUseCase",
     "PenaAccessDeniedError",
+    "PenaAccountabilityAccessDeniedError",
+    "PenaAccountabilityExpenseNotFoundError",
+    "PenaAccountabilityMemberNotFoundError",
+    "PenaAccountabilityPenaNotFoundError",
     "PenaLabelsAccessDeniedError",
     "PenaLabelsPenaNotFoundError",
     "PenaProfileAccessDeniedError",
